Replaced Router Modules with Global Prefix, reorganized file structure
