+++ .gitignore
@@ -34,5 +34,4 @@ lerna-debug.log*
 !.vscode/launch.json
 !.vscode/extensions.json
 
-/db
 .env
