+++ src/entities/index.ts
@@ -1,0 +1,3 @@
+import { Season } from './season.entity'
+
+export default [ Season ];
