+++ src/entities/season.entity.ts
