--- src/entity/api/api.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller()
-export class ApiController {}
