+++ src/modules/index.ts
@@ -1,0 +1,5 @@
+import { SeasonsModule } from "./seasons/seasons.module";
+
+export const Modules = [
+    SeasonsModule
+]
