+++ src/modules/seasons/seasons.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get } from '@nestjs/common';
 import { SeasonsService } from './seasons.service';
 
-@Controller()
+@Controller('seasons')
 export class SeasonsController {
     constructor(private readonly seasonsService: SeasonsService) {}
